+++ api/src/school/infrastructure/repositories/mission-learner-repository.js
@@ -1,14 +1,9 @@
-import { knex } from '../../../../db/knex-database-connection.js';
-import { fetchPage } from '../../../shared/infrastructure/utils/knex-utils.js';
 import { MissionLearner } from '../../domain/models/MissionLearner.js';
 
-const findPaginatedMissionLearners = async function ({ organizationId, page }) {
-  const { results, pagination } = await fetchPage(
-    knex('organization-learners').select('*').where({ organizationId: organizationId }),
-    page,
-  );
+const findPaginatedMissionLearners = async function ({ organizationId, page, organizationLearnerApi }) {
+  const { organizationLearners, pagination } = await organizationLearnerApi.find({ organizationId, page });
 
-  const missionLearners = results.map((missionLearner) => new MissionLearner({ ...missionLearner }));
+  const missionLearners = organizationLearners.map((missionLearner) => new MissionLearner({ ...missionLearner }));
   return { missionLearners, pagination };
 };
